requirements.txt
tensorflow-gpu==2.5.0
pillow==8.3.2
scipy